chore(zui): extract zod error logic out of types
