+++ packages/zui/src/z/error/index.ts
@@ -1,15 +1,6 @@
-import { TypeOf, ZodType, ZodParsedType, util, Primitive } from '../index'
+import { util, ZodParsedType } from '../types/utils'
+import { Primitive } from '../types/utils/typeAliases'
 import { errorMap as defaultErrorMap } from './locales/en'
-
-type allKeys<T> = T extends any ? keyof T : never
-
-export type inferFlattenedErrors<T extends ZodType<any, any, any>, U = string> = typeToFlattenedError<TypeOf<T>, U>
-export type typeToFlattenedError<T, U = string> = {
-  formErrors: U[]
-  fieldErrors: {
-    [P in allKeys<T>]?: U[]
-  }
-}
 
 export const ZodIssueCode = util.arrayToEnum([
   'invalid_type',
@@ -186,8 +177,6 @@ export type ZodFormattedError<T, U = string> = {
   _errors: U[]
 } & recursiveZodFormattedError<NonNullable<T>>
 
-export type inferFormattedError<T extends ZodType<any, any, any>, U = string> = ZodFormattedError<TypeOf<T>, U>
-
 export class ZodError<T = any> extends Error {
   issues: ZodIssue[] = []
 
@@ -287,26 +276,6 @@ export class ZodError<T = any> extends Error {
 
   addIssues = (subs: ZodIssue[] = []) => {
     this.issues = [...this.issues, ...subs]
-  }
-
-  flatten(): typeToFlattenedError<T>
-  flatten<U>(mapper?: (issue: ZodIssue) => U): typeToFlattenedError<T, U>
-  flatten<U = string>(mapper: (issue: ZodIssue) => U = (issue: ZodIssue) => issue.message as U): any {
-    const fieldErrors: any = {}
-    const formErrors: U[] = []
-    for (const sub of this.issues) {
-      if (sub.path.length > 0) {
-        fieldErrors[sub.path[0]!] = fieldErrors[sub.path[0]!] || []
-        fieldErrors[sub.path[0]!].push(mapper(sub))
-      } else {
-        formErrors.push(mapper(sub))
-      }
-    }
-    return { formErrors, fieldErrors }
-  }
-
-  get formErrors() {
-    return this.flatten()
   }
 }
