+++ packages/zui/src/z/types/array/index.ts
@@ -1,6 +1,6 @@
 import { isEqual } from 'lodash-es'
+import { ZodIssueCode } from '../../error'
 import {
-  ZodIssueCode,
   ParseInputLazyPath,
   RawCreateParams,
   ZodType,
