+++ packages/zui/src/z/types/catch/index.ts
@@ -1,5 +1,5 @@
+import { ZodError } from '../../error'
 import {
-  ZodError,
   RawCreateParams,
   ZodType,
   ZodTypeDef,
