+++ packages/zui/src/z/types/discriminatedUnion/index.ts
@@ -1,10 +1,10 @@
+import { ZodIssueCode } from '../../error'
 import { unique } from '../../utils'
 import {
   ZodBranded,
   ZodCatch,
   ZodDefault,
   ZodEnum,
-  ZodIssueCode,
   input,
   output,
   RawCreateParams,
