+++ packages/zui/src/z/types/function/index.ts
@@ -1,11 +1,6 @@
+import { defaultErrorMap, getErrorMap, ZodError, ZodErrorMap, ZodIssue, ZodIssueCode } from '../../error'
 import { unique } from '../../utils'
 import {
-  defaultErrorMap,
-  getErrorMap,
-  ZodError,
-  ZodErrorMap,
-  ZodIssue,
-  ZodIssueCode,
   RawCreateParams,
   ZodType,
   ZodTypeDef,
