+++ packages/zui/src/z/types/index.ts
@@ -22,7 +22,6 @@ export * from './date'
 export * from './default'
 export * from './discriminatedUnion'
 export * from './enum'
-export * from './error'
 export * from './function'
 export * from './index'
 export * from './intersection'
@@ -51,6 +50,3 @@ export * from './undefined'
 export * from './union'
 export * from './unknown'
 export * from './void'
-
-import defaultErrorMap from './error/locales/en'
-export { defaultErrorMap }
