+++ packages/zui/src/z/types/literal/index.ts
@@ -1,7 +1,7 @@
 import { isEqual } from 'lodash-es'
 
+import { ZodIssueCode } from '../../error'
 import {
-  ZodIssueCode,
   RawCreateParams,
   ZodType,
   ZodTypeDef,
