+++ packages/zui/src/z/types/map/index.ts
@@ -1,6 +1,6 @@
+import { ZodIssueCode } from '../../error'
 import { unique } from '../../utils'
 import {
-  ZodIssueCode,
   ParseInputLazyPath,
   RawCreateParams,
   ZodType,
