+++ packages/zui/src/z/types/nativeEnum/index.ts
@@ -1,7 +1,7 @@
 import { isEqual } from 'lodash-es'
 
+import { ZodIssueCode } from '../../error'
 import {
-  ZodIssueCode,
   RawCreateParams,
   ZodType,
   ZodTypeDef,
