+++ packages/zui/src/z/types/record/index.ts
@@ -1,7 +1,7 @@
+import { ZodIssueCode } from '../../error'
 import { unique } from '../../utils'
 import {
   BRAND,
-  ZodIssueCode,
   ParseInputLazyPath,
   RawCreateParams,
   ZodType,
