+++ packages/zui/src/z/types/set/index.ts
@@ -1,5 +1,5 @@
+import { ZodIssueCode } from '../../error'
 import {
-  ZodIssueCode,
   ParseInputLazyPath,
   RawCreateParams,
   ZodType,
